+++ pomcorn/page.py
@@ -175,6 +175,12 @@ class Page(WebView):
             relative_url (str): Relative URL
 
         """
-        if not relative_url.startswith("/"):
-            relative_url = f"/{relative_url}"
-        return f"{app_root}{relative_url}"
+        if app_root.endswith("/"):
+            # https://www.youtube.com/ -> https://www.youtube.com
+            app_root = app_root[:-1]
+
+        if relative_url.startswith("/"):
+            # /watch_list -> watch_list
+            relative_url = relative_url[1:]
+
+        return f"{app_root}/{relative_url}"