+++ docs/CHANGELOG.rst
@@ -6,6 +6,7 @@ We follow `Semantic Versions <https://semver.org/>`_.
 0.7.2
 *******************************************************************************
 - Improve ``Page.click_on_page`` method to click on <html> tag
+- Improve ``Page.open_from_url`` to support kwargs
 
 
 0.7.1